+++ packages/ui/src/components/Dropdown/Dropdown.stories.tsx
@@ -8,6 +8,7 @@ export default { component: Dropdown } satisfies Meta<typeof Dropdown>;
 
 export const Default: Story = {
   args: {
+    onSelection: (option) => alert(option),
     options: ['Option 1', 'Option 2'],
     title: 'Dropdown'
   }