+++ packages/ui/src/components/Dropdown/Dropdown.tsx
@@ -1,9 +1,6 @@
-import { Menu, Transition } from '@headlessui/react';
-import { twMerge } from 'tailwind-merge';
+import { ChevronDownIcon } from '@radix-ui/react-icons';
 
-import { type ButtonProps } from '../Button';
-import { Card } from '../Card/Card';
-import { DropdownButton } from '../DropdownButton';
+import { Select } from '../Select';
 
 type DropdownOptions = Record<string, string> | readonly string[];
 
@@ -22,63 +19,32 @@ export type DropdownProps<T extends DropdownOptions> = {
   /** Either a list of options for the dropdown, or an object with options mapped to custom labels  */
   options: T;
 
-  size?: 'lg' | 'md' | 'sm';
-
   /** The text content for the dropdown toggle */
   title: string;
-
-  /** The button variant to use for the dropdown toggle */
-  variant?: ButtonProps['variant'];
 };
 
 // eslint-disable-next-line react/function-component-definition
 export function Dropdown<const T extends DropdownOptions>({
-  className,
   onSelection,
   options,
-  size,
-  title,
-  variant
+  title
 }: DropdownProps<T>) {
   const optionKeys: readonly string[] = options instanceof Array ? options : Object.keys(options);
   return (
-    <Menu as="div" className={twMerge('relative w-full whitespace-nowrap', className)}>
-      <Menu.Button
-        as={DropdownButton}
-        className="h-full w-full"
-        disabled={options.length === 0}
-        size={size}
-        style={{ width: '100%' }}
-        variant={variant}
-      >
+    <Select value="" onValueChange={onSelection}>
+      <Select.Trigger>
         {title}
-      </Menu.Button>
-      <Transition
-        as="div"
-        className="absolute bottom-0 z-10 w-full"
-        enter="transition ease-out duration-100"
-        enterFrom="transform opacity-0 scale-95"
-        enterTo="transform opacity-100 scale-100"
-        leave="transition ease-in duration-75"
-        leaveFrom="transform opacity-100 scale-100"
-        leaveTo="transform opacity-0 scale-95"
-      >
-        <Menu.Items as={Card} className="absolute z-10 mt-2 flex w-fit min-w-full flex-col overflow-hidden">
+        <ChevronDownIcon />
+      </Select.Trigger>
+      <Select.Content>
+        <Select.Group>
           {optionKeys.map((option) => (
-            <Menu.Item key={option}>
-              <button
-                className="p-2 text-left hover:bg-slate-200 dark:hover:bg-slate-700"
-                style={{ minWidth: 100 }}
-                onClick={() => {
-                  onSelection(option as DropdownOptionKey<T>);
-                }}
-              >
-                {Array.isArray(options) ? option : (options[option as keyof T] as string)}
-              </button>
-            </Menu.Item>
+            <Select.Item key={option} value={option}>
+              {Array.isArray(options) ? option : (options[option as keyof T] as string)}
+            </Select.Item>
           ))}
-        </Menu.Items>
-      </Transition>
-    </Menu>
+        </Select.Group>
+      </Select.Content>
+    </Select>
   );
 }