+++ packages/ui/src/components/Select/SelectTrigger.tsx
@@ -9,7 +9,14 @@ export const SelectTrigger = React.forwardRef<
   React.ComponentPropsWithoutRef<typeof SelectPrimitive.Trigger>
 >(function SelectTrigger({ children, className, ...props }, ref) {
   return (
-    <SelectPrimitive.Trigger className={cn('select-trigger', className)} ref={ref} {...props}>
+    <SelectPrimitive.Trigger
+      className={cn(
+        'flex h-9 w-full items-center justify-between whitespace-nowrap rounded-md border border-input bg-transparent px-3 py-2 text-sm shadow-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring disabled:cursor-not-allowed disabled:opacity-50 [&>span]:line-clamp-1',
+        className
+      )}
+      ref={ref}
+      {...props}
+    >
       {children}
     </SelectPrimitive.Trigger>
   );