+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Button: Removed `label` and `icon` from props, can be replaced by `children`
 - Tailwind: Removed `text-muted` and `text-strong`, can be replaced by `text-muted-foreground` and `text-primary-foreground` respectively
 - `FormPageWrapper`: Deleted
+- Dropdown: Remove `className` and `size` props
 
 ### Minor Changes
 - ArrowToggle: Added background effect on hover